refactor: fix long tests
